packages/pqcow_client/pyproject.toml
[project]
name = "pqcow-client"
version = "0.0.1"
dependencies = [
    "liboqs-python==0.10.0",
    "msgspec==0.18.6",
    "cryptography==43.0.3",
    "pqcow-types",
]

[tool.uv]
package = true
